+++ src/modules/core/components/UploadAddresses/index.ts
@@ -1,0 +1,1 @@
+export { default } from './UploadAddresses';
