+++ src/modules/dashboard/components/Dialogs/ManageWhitelistDialog/AddAddressesTab.css.d.ts
