+++ src/modules/dashboard/components/Dialogs/ManageWhitelistDialog/ManageWhitelistDialog.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback } from 'react';
+import React, { useCallback, useState } from 'react';
 import { FormikProps } from 'formik';
 import * as yup from 'yup';
 import { useHistory } from 'react-router-dom';
@@ -9,6 +9,11 @@ import { ActionForm } from '~core/Fields';
 import { Colony } from '~data/index';
 import { ActionTypes } from '~redux/index';
 import { WizardDialogType } from '~utils/hooks';
+import {
+  mergeSchemas,
+  validationSchemaInput,
+  validationSchemaFile,
+} from '~utils/whitelistValidation';
 import { pipe, withMeta, mapPayload } from '~utils/actions';
 import { Address } from '~types/index';
 
@@ -35,20 +40,41 @@ const ManageWhitelistDialog = ({
   callStep,
   prevStep,
   colony,
+  colony: { colonyAddress },
 }: Props) => {
+  const [showInput, setShowInput] = useState<boolean>(true);
+  const toggleShowInput = () => setShowInput(!showInput);
+
   const history = useHistory();
 
   const validationSchema = yup.object().shape({
     annotation: yup.string().max(4000),
   });
 
+  const mergedSchemas = mergeSchemas(
+    validationSchema,
+    showInput ? validationSchemaInput : validationSchemaFile,
+  );
+
   const transform = useCallback(
     pipe(
-      mapPayload(({ annotation: annotationMessage }) => {
-        return {
-          annotationMessage,
-        };
-      }),
+      mapPayload(
+        ({
+          annotation: annotationMessage,
+          userAddresses: whitelistAddress,
+          whitelistCSVUploader,
+        }) => {
+          return {
+            annotationMessage,
+            userAddresses:
+              whitelistAddress !== undefined
+                ? [whitelistAddress]
+                : whitelistCSVUploader[0].parsedData,
+            colonyAddress,
+            status: true,
+          };
+        },
+      ),
       withMeta({ history }),
     ),
     [],
@@ -79,11 +105,12 @@ const ManageWhitelistDialog = ({
         whitelistedAddresses: whitelistedUsers.map(
           (user) => user.profile.walletAddress,
         ),
+        isSubmitting: false,
       }}
       submit={ActionTypes.COLONY_ACTION_GENERIC}
       error={ActionTypes.COLONY_ACTION_GENERIC_ERROR}
       success={ActionTypes.COLONY_ACTION_GENERIC_SUCCESS}
-      validationSchema={validationSchema}
+      validationSchema={mergedSchemas}
       onSuccess={close}
       transform={transform}
     >
@@ -94,6 +121,8 @@ const ManageWhitelistDialog = ({
             colony={colony}
             whitelistedUsers={whitelistedUsers}
             back={() => callStep(prevStep)}
+            showInput={showInput}
+            toggleShowInput={toggleShowInput}
           />
         </Dialog>
       )}
