--- src/modules/dashboard/components/Whitelist/UploadAddressesWidget/index.ts
@@ -1,3 +1,0 @@
-export { default } from './UploadAddressesWidget';
-
-export { default as KYCUploadAddressesWidget } from './KYCUploadAddressesWidget';
