+++ src/modules/dashboard/components/Whitelist/WhitelistAddresses/KYCUploadAddressesWidget.css
@@ -1,0 +1,53 @@
+.container {
+  margin: 40px 0;
+}
+
+.validationError {
+  font-size: var(--size-tiny);
+  font-weight: var(--weight-bold);
+  line-height: 18px;
+  color: var(--danger);
+}
+
+.inputContainer {
+  margin-bottom: 9px;
+}
+
+.actionsContainer {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-bottom: 6px;
+  width: unset;
+}
+
+.actionsContainer > label {
+  margin-bottom: 0px;
+}
+
+.actionsSubContainer {
+  display: flex;
+  align-items: center;
+}
+
+.actionsSubContainer > label {
+  margin-right: 0px;
+}
+
+.buttonsContainer {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+}
+
+.agreeemntButton {
+  margin-right: 30px;
+}
+
+.uploaderContainer {
+  margin-bottom: 28px;
+}
+
+.uploaderContainer > div > p {
+  height: 0;
+}
