+++ src/modules/dashboard/components/Whitelist/WhitelistAddresses/KYCUploadAddressesWidget.css.d.ts
@@ -1,0 +1,8 @@
+export const container: string;
+export const validationError: string;
+export const inputContainer: string;
+export const actionsContainer: string;
+export const actionsSubContainer: string;
+export const buttonsContainer: string;
+export const agreeemntButton: string;
+export const uploaderContainer: string;
